refactor: ディレクトリ構造とインポートパスを整理し、不要なツールを削除
