+++ package/main/src/Date/index.ts
@@ -1,4 +1,6 @@
+export * from "./birthday";
 export * from "./dateRange";
+export * from "./dayOfWeek";
 export * from "./getDay";
 export * from "./isLeapYear";
 export * from "./new";
