+++ package/main/src/Simple/Date/index.ts
@@ -1,1 +1,3 @@
+export * from "./birthdaySimple";
+export * from "./dayOfWeekSimple";
 export * from "./nowSimple";
