+++ package/main/src/Simple/Math/deviationValueSimple.ts
