+++ package/main/src/Simple/Math/index.ts
@@ -1,0 +1,1 @@
+export * from "./deviationValueSimple";
