--- package/main/src/Simple/Tool/index.ts
@@ -1,3 +1,0 @@
-export * from "./birthdaySimple";
-export * from "./dayOfWeekSimple";
-export * from "./deviationValueSimple";
