+++ package/main/src/Simple/index.ts
@@ -1,3 +1,3 @@
 export * from "./Array";
 export * from "./Date";
-export * from "./Tool";
+export * from "./Math";
