+++ package/main/src/String/deleteSpace.ts
