+++ package/main/src/String/index.ts
@@ -1,3 +1,4 @@
+export * from "./deleteSpace";
 export * from "./formatString";
 export * from "./fromBase64";
 export * from "./padEnd";
