+++ package/main/src/Tool/index.ts
@@ -1,10 +1,1 @@
-export * from "./birthday";
-export * from "./cmykToRgba";
 export * from "./createPipeline";
-export * from "./dayOfWeek";
-export * from "./deleteSpace";
-export * from "./hexaToRgba";
-export * from "./hslaToRgba";
-export * from "./rgbaToCmyk";
-export * from "./rgbaToHexA";
-export * from "./rgbaToHsla";
